Removed more array dependencies
